feat: modified structure of sdb in run user
