+++ run_user/index.js
@@ -2,6 +2,7 @@
 const config          = require('../config.js');
 const default_headers = require('../comum/default_headers.js');
 const request         = require('../comum/tago_request.js');
+const SDB             = require('./sdb.js');
 
 class RunUser {
   constructor(token) {
@@ -168,48 +169,9 @@ class RunUser {
     return request(options);
   }
 
-  /**
-   * Retrieves a custom parameter of a run user.
-   * The run user is identified by the token in the constructor.
-   * @param {String} tagoRunURL - the url of tago run.
-   * @param {String} key - identifier of the parameter.
-   */
-  paramGet(tagoRunURL, key) {
-    const url    = `${config.api_url}/run/${tagoRunURL}/sdb/${key}`;
-    const method = 'GET';
-
-    const options = { ...this.default_options, url, method };
-    return request(options);
-  }
-
-  /**
-   * Creates or updates a custom parameter of a run user.
-   * The run user is identified by the token in the constructor.
-   * @param {String} tagoRunURL - the url of tago run.
-   * @param {String} key - identifier of the parameter.
-   * @param {String} value - value of the parameter.
-   */
-  paramSet(tagoRunURL, key, value) {
-    const data   = { value };
-    const url    = `${config.api_url}/run/${tagoRunURL}/sdb/${key}`;
-    const method = 'POST';
-
-    const options = { ...this.default_options, url, method, data };
-    return request(options);
-  }
-
-  /**
-   * Delete a custom parameter of a run user.
-   * The run user is identified by the token in the constructor.
-   * @param {String} tagoRunURL - the url of tago run.
-   * @param {String} key - identifier of the parameter.
-   */
-  paramDelete(tagoRunURL, key) {
-    const url    = `${config.api_url}/run/${tagoRunURL}/sdb/${key}`;
-    const method = 'DELETE';
-
-    const options = { ...this.default_options, url, method };
-    return request(options);
+  // ----------- Sub-methods -----------
+  get SDB() {
+    return new SDB(this.token);
   }
 }
